Cast it as a Long and then to an Integer

+++ platforms/common-kord/src/main/kotlin/net/perfectdreams/discordinteraktions/platforms/kord/commands/KordCommandRegistry.kt
@@ -163,7 +163,7 @@ class KordCommandRegistry(private val applicationId: Snowflake, private val rest
                     this.required = !cmdOption.type.isNullable
 
                     for (choice in cmdOption.choices) {
-                        choice(choice.name, choice.value as Int)
+                        choice(choice.name, (choice.value as Long).toInt())
                     }
                 }
             CommandOptionType.Number, CommandOptionType.NullableNumber ->
